Day14/requirements.txt
groq
sentence-transformers
faiss-cpu
python-dotenv

